+++ randovania/game_connection/executor/nintendont_executor.py
@@ -99,6 +99,8 @@ class NintendontExecutor(MemoryOperationExecutor):
     _socket: SocketHolder | None = None
     _socket_error: Exception | None = None
 
+    SUPPORTED_API_VERSION = 1
+
     _timeout = 5
     # timeout in seconds on when we disconnect when we don't get a response.
     # The Prime games (which this is currently used for) are somewhat time-sensitive. If we can't ensure a timely
@@ -148,7 +150,23 @@ class NintendontExecutor(MemoryOperationExecutor):
 
             self.logger.debug("Waiting for API details response.")
             response = await asyncio.wait_for(reader.read(1024), timeout=self._timeout)
-            api_version, max_input, max_output, max_addresses = struct.unpack_from(">IIII", response, 0)
+            invalid_message = f"Unable to connect to {self._ip}:{self._port} - Unsupported Nintendont version!"
+            try:
+                api_version, max_input, max_output, max_addresses = struct.unpack(">IIII", response)
+
+            except struct.error as e:
+                writer.close()
+                self._socket_error = e
+                return f"{invalid_message} Invalid response when requesting API details."
+
+            if api_version != self.SUPPORTED_API_VERSION:
+                writer.close()
+                return f"{invalid_message} Nintendont has API {api_version} but expected {self.SUPPORTED_API_VERSION}."
+
+            if max_input > 256 or max_output > 256 or max_addresses > 16:
+                # 256, 256 and 16 are the current theoretical maximum values for input/output/address per the protocol.
+                writer.close()
+                return f"{invalid_message} Nintendont responding with invalid API details."
 
             self.logger.debug(f"Remote replied with API level {api_version}, connection successful.")
             self._socket = SocketHolder(reader, writer, api_version, max_input, max_output, max_addresses)
@@ -163,7 +181,7 @@ class NintendontExecutor(MemoryOperationExecutor):
                 raise e
 
             return "Currently in the Homebrew Channel. Upload Nintendont or launch it manually."
-        except (TimeoutError, OSError, struct.error, UnicodeError) as e:
+        except (TimeoutError, OSError, UnicodeError) as e:
             # UnicodeError is for some invalid ip addresses
             self._socket = None
             message = f"Unable to connect to {self._ip}:{self._port} - ({type(e).__name__}) {e}"